+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 2.2
+
+#### Changes
+
+#### Fixes
+
+* Node: Resolved circular import dependency between `BaseClient.ts`, `GlideClient.ts`, and `GlideClusterClient.ts.` ([#4607](https://github.com/valkey-io/valkey-glide/issues/4607))
+
+#### Operational Enhancements
+
 ## 2.1
 
 #### Changes